backend/requirements.txt
Flask==2.3.2
tensorflow==2.15.0
numpy==1.26.4
Pillow==10.3.0
joblib==1.4.2
scikit-learn==1.5.0
gunicorn==22.0.0
python-multipart==0.0.9